+++ frontend/src/api/auth.js
@@ -1,7 +1,7 @@
 import { API_URL } from "../config";
 import { api } from "../utils/api";
 
-export const fetchUsers = async () => {
+export const fetchUser = async () => {
   return api
     .get(`${API_URL}/api/users`)
     .then(({ data }) => {
