+++ frontend/src/components/Footer.jsx
@@ -1,7 +1,7 @@
 const Footer = () => {
 	console.log('Not implemented');
 	return (
-		<div className="bg-primary-700 bottom-0 
+		<div className="bg-primary-700 bottom-0 text-tc
         md:px-5 md:py-10 md:mt-12
         pt-5 w-screen z-20">
 			<span>Group 8</span>
