+++ frontend/src/components/Navbar.jsx
@@ -2,7 +2,7 @@ import Searchbar from './Searchbar';
 
 const Navbar = () => {
 	return (
-		<div className='fixed bg-primary-700 top-0 
+		<div className='fixed bg-primary-700 top-0 text-tc
 			md:px-5 md:py-10
 			pt-5 w-screen z-20'>
 			<h1 className='text-center' id='nav-header'>IV1201</h1>
