+++ frontend/src/layouts/index.js
@@ -1,0 +1,3 @@
+export { default as AuthLayout } from './AuthLayout';
+export { default as HomeLayout } from './HomeLayout';
+export { default as ProtectedLayout } from './ProtectedLayout';
