+++ frontend/src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as ApplicantsPage } from './ApplicantsPage';
+export { default as HomePage } from './HomePage';
+export { default as LoginPage } from './LoginPage';
